Work round Quarto parsing
In answer to the questions in the previous commit message:

a) It turns out that Quarto is doing some preprocessing before parsing,
that can't easily be replicated with Pandoc - see:
https://github.com/quarto-dev/quarto-cli/discussions/11393.  Work round
with a fragile fix for our known test cases, using a regex.

b) The previous code wasn't testing this parsing, because I was
modifying the notebooks myself - with the correct backtick code blocks -
rather than looking at the generated notebooks.